feat: 忘了（梅开n度

+++ src/main/kotlin/cn/spacexc/bilibilisdk/sdk/user/history/remote/Cursor.kt
@@ -2,7 +2,7 @@ package cn.spacexc.bilibilisdk.sdk.user.history.remote
 
 data class Cursor(
     val business: String,
-    val max: Int,
+    val max: Long,
     val ps: Int,
     val view_at: Int
 )
